VSL-120: Moved sent from and sent to to a separate component
